Add files via upload
This work aims to identify the determinants of load generation in the Brazilian electricity system, based on time series modelling of different levels of variables, such as: macroeconomic, institutional and technological progress.
The central idea of this work is to design an essay to discuss these determinants, to establish possible models that can explain them and, finally, to guide more robust research.
The paper is divided into three sections as follows:
The first one discusses the literature review of the sector and the determinants that have already been identified over time. In this way, the review aims to raise possible hypotheses and conclusions already published in order to enrich the debate within the proposed project. 
The second section deals with data and methodology. In this sense, the data sources, tests, guidelines for the selection of variables and, consequently, the proposed tests and models are presented.
Finally, the conclusion focuses on the comparison between the initial hypothesis of the study and the results obtained by the models. There is also an explanation of the limitations of the study, as well as suggestions that could form part of a new round of exploratory analysis of the topic now established.